Reduced background flickering
The background brightness changes are now changing to a set number of levels, rather than a +/- 1 every update interval. This makes the background change in a slightly less jarring way.

Also fixed a possible bug where the background and highlight brightness would go over 127 and effectively overflow the max brightness that the strip can deliver, causing for unintended flickering as the output goes near-dark momentarily.

Added a good healthy amount of comments explaining the function of some of the constants that can be defined.